+++ bogenliga/bogenliga-application/src/main/java/de/bogenliga/application/services/v1/download/DownloadService.java
@@ -20,10 +20,10 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
-import de.bogenliga.application.business.Bogenkontrollliste.api.BogenkontrolllisteComponent;
-import de.bogenliga.application.business.Meldezettel.api.MeldezettelComponent;
-import de.bogenliga.application.business.Schusszettel.api.SchusszettelComponent;
-import de.bogenliga.application.business.Setzliste.api.SetzlisteComponent;
+import de.bogenliga.application.business.bogenkontrollliste.api.BogenkontrolllisteComponent;
+import de.bogenliga.application.business.meldezettel.api.MeldezettelComponent;
+import de.bogenliga.application.business.schusszettel.api.SchusszettelComponent;
+import de.bogenliga.application.business.setzliste.api.SetzlisteComponent;
 import de.bogenliga.application.business.lizenz.api.LizenzComponent;
 import de.bogenliga.application.common.errorhandling.ErrorCode;
 import de.bogenliga.application.common.errorhandling.exception.TechnicalException;
