--- bogenliga/bogenliga-business-logic/src/main/java/de/bogenliga/application/business/Meldezettel/api/MeldezettelComponent.java
@@ -1,17 +1,0 @@
-package de.bogenliga.application.business.Meldezettel.api;
-
-import de.bogenliga.application.common.component.ComponentFacade;
-
-/**
- * Responsible for the meldezettel database requests.
- */
-public interface MeldezettelComponent extends ComponentFacade {
-
-    /**
-     * Generates a pdf as binary document
-     * @param wettkampfid ID for the competition
-     * @return document
-     */
-    byte[] getMeldezettelPDFasByteArray(long wettkampfid);
-
-}
