+++ bogenliga/bogenliga-business-logic/src/main/java/de/bogenliga/application/business/setzliste/api/types/SetzlisteDO.java
@@ -1,0 +1,13 @@
+package de.bogenliga.application.business.setzliste.api.types;
+
+import de.bogenliga.application.common.component.types.CommonDataObject;
+
+/**
+ * TODO [AL] class documentation
+ *
+ * @author Andre Lehnert, eXXcellent solutions consulting & software gmbh
+ */
+public class SetzlisteDO extends CommonDataObject {
+    private static final long serialVersionUID = 3872635461141356788L;
+
+}
